requirements.txt
pandas_stubs==1.2.0.31
numpy==1.23.4
matplotlib==3.2.0
wandb==0.13.3
opencv_python==4.5.3.56
scipy==1.5.3
torch==1.12.1
torchvision==0.13.1
tqdm==4.62.3
requests==2.26.0
albumentations==1.3.0
coremltools==6.1
onnx==1.12.0
onnx_graphsurgeon==0.0.1.dev5
onnxsim==0.4.10
pafy==0.5.5
pandas==1.5.1
Pillow==9.3.0
pycocotools==2.0.6
PyYAML==6.0
seaborn==0.12.1
thop==0.1.1.post2209072238
tritonclient==2.27.0
